feat: move actionType to textX
