+++ packages/core/src/docs/data-model/text-x/apply.ts
@@ -19,7 +19,7 @@ import type { Nullable } from '../../../shared';
 import { Tools } from '../../../shared';
 import { UpdateDocsAttributeType } from '../../../shared/command-enum';
 import type { IDocumentBody } from '../../../types/interfaces';
-import { type TextXAction, TextXActionType } from '../action-types';
+import { type TextXAction, TextXActionType } from './action-types';
 import { updateAttributeByDelete } from './apply-utils/delete-apply';
 import { updateAttributeByInsert } from './apply-utils/insert-apply';
 import { updateAttribute } from './apply-utils/update-apply';
