+++ packages/core/src/index.ts
@@ -43,7 +43,7 @@ export {
     type IDeleteAction,
     type IInsertAction,
     type IRetainAction,
-} from './docs/data-model/action-types';
+} from './docs/data-model/text-x/action-types';
 export { DataValidationRenderMode } from './types/enum/data-validation-render-mode';
 export { ActionIterator } from './docs/data-model/text-x/action-iterator';
 export { getBodySlice, composeBody } from './docs/data-model/text-x/utils';
